Initial commit
This demonstrates an improvement that could be made to rules_go to
identify cases where the build context prevents any sources from
being compiled in a go_library target. Currently, the build for
that target will succeed, but any go_binary or go_test targets
that depend on the go_library target will fail with unintuitive
error messages.